Former-commit-id: b7d3089f46f7b70b033e0a737c738811e7718f8d [formerly 5d1d3d47913baac8731ef054c70daf561903c71c]
Former-commit-id: 204dc0d753d683bd9ced39de71b19c5a17cfe101

+++ kraken/lib/src/bridge/from_native.dart
@@ -21,7 +21,7 @@ import 'native_types.dart';
 class NativeString extends Struct {
   Pointer<Uint16> string;
 
-  @Int64()
+  @Int32()
   int length;
 }
